Moved prediction stuff to prediction component
